requirements.txt
streamlit
pillow
pandas
scikit-learn
numpy
matplotlib
lightgbm